requirements.txt
Flask==1.0.2
Flask-RESTful==0.3.6
Werkzeug==0.14.1
mysql-connector==2.2.9
